Add a temporary shim to bridge Go 1.23 and Go 1.24 (#2)
Prior to Go 1.24, the cacheprog request message used "ObjectID" in its "put"
request message to the plugin. For the non-experiment release in Go 1.24, it is
being changed to "OutputID" to match the response message.

As a workaround in the meantime, make the server handle both names, preferring
the newer one. The real toolchain will never set both, so this should be fine.

Fixes #1